norminett & added ft_itoa(s) to global & more
