+++ parsing/big_list.c
@@ -6,7 +6,7 @@
 /*   By: mkaruvan <mkaruvan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/04 09:49:19 by mkaruvan          #+#    #+#             */
-/*   Updated: 2022/06/27 09:57:35 by mkaruvan         ###   ########.fr       */
+/*   Updated: 2022/07/01 14:27:34 by mkaruvan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -79,36 +79,4 @@ void	lst_big_new(t_new **cmd, t_list *lst)
 		(*cmd)->prev = temp;
 		lst = lst->next;
 	}
-}
-
-void	lst_add(t_new **cmd, t_list *lst)
-{
-	if ((*cmd)->prev == NULL)
-		lst_big_new(cmd, lst);
-	else
-		lst_add_front(cmd, lst);
-}
-
-void	lst_skip_node2(t_new *cmd)
-{
-	if (cmd->prev)
-		cmd->prev->next = cmd->next;
-	if (cmd->next)
-		cmd->next->prev = cmd->prev;
-}
-
-void	make_big_list(t_new **cmd)
-{
-	t_new	*temp;
-
-	temp = (*cmd);
-	while ((*cmd) != NULL)
-	{
-		if ((*cmd)->lst != NULL)
-			big_list_help(cmd);
-		(*cmd) = (*cmd)->next;
-	}
-	*cmd = temp;
-	while ((*cmd)->prev)
-		*cmd = (*cmd)->prev;
 }
