+++ parsing/p_dollar.c
@@ -6,7 +6,7 @@
 /*   By: mkaruvan <mkaruvan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/27 15:36:54 by mkaruvan          #+#    #+#             */
-/*   Updated: 2022/06/27 10:13:10 by mkaruvan         ###   ########.fr       */
+/*   Updated: 2022/07/01 15:23:49 by mkaruvan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
